+++ src/ska_sdp_global_sky_model/api/app/request_responder.py
@@ -106,6 +106,8 @@ class QueryParameters:
         """
         # Query components within the field of view using spatial index
         # pylint: disable=no-member,duplicate-code
+        metadata_records = self._get_metadata_records(db)
+        ids = [record.id for record in metadata_records]
         sky_components_query = (
             db.query(SkyComponent)
             .where(
@@ -117,12 +119,21 @@ class QueryParameters:
                     self.fov_deg,
                 )
             )
-            .where(SkyComponent.version == self.get_version(db))
-            .where(SkyComponent.catalogue_name == self.catalogue_name)
+            .where(SkyComponent.gsm_id.in_(ids))
         )
         query_builder = QueryBuilder(SkyComponent, self.component_queries)
         sky_components_query = query_builder.apply_filters(sky_components_query)
-        return sky_components_query.all()
+        return metadata_records, sky_components_query.all()
+
+    def _get_metadata_records(self, db) -> list[GlobalSkyModelMetadata]:
+        metadata_query = (
+            db.query(GlobalSkyModelMetadata)
+            .where(GlobalSkyModelMetadata.version == self.get_version(db))
+            .where(GlobalSkyModelMetadata.catalogue_name == self.catalogue_name)
+        )
+        query_builder = QueryBuilder(GlobalSkyModelMetadata, self.metadata_queries)
+        metadata_query = query_builder.apply_filters(metadata_query)
+        return metadata_query.all()
 
     def get_version(self, db: Session) -> str:
         """Get version. If was 'latest' set convert to latest semantic version in database.
@@ -255,10 +266,12 @@ def _get_flows(txn: ska_sdp_config.Config.txn) -> Generator[(Flow, FlowSource)]:
         if status in ["COMPLETED", "FAILED"]:
             continue
 
-        expected_state = "INITIALISED" if resource_toggle.is_active() else "PENDING"
-        if status != expected_state:
+        expected_states = ["INITIALISED"]
+        if not resource_toggle.is_active():
+            expected_states += ["PENDING"]
+        if status not in expected_states:
             logger.debug(
-                "%s -> not in correct state %s != %s", key, state.get("status"), expected_state
+                "%s -> not in correct state %s != %s", key, state.get("status"), expected_states
             )
             continue
 
@@ -337,24 +350,19 @@ def _query_gsm_for_lsm(
 
         # Query metadata for this version
 
-        metadata_row = (
-            db.query(GlobalSkyModelMetadata)
-            .filter(GlobalSkyModelMetadata.version == resolved_version)
-            .first()
-        )
-        metadata_dict = metadata_row.columns_to_dict() if metadata_row else {}
-
         sky_components_dict = {}
-        for sky_component in query_parameters.sky_components(db):
+        metadata_records, sky_components = query_parameters.sky_components(db)
+        for sky_component in sky_components:
             sky_component_dict = sky_component.columns_to_dict()
             # Remove database-specific fields that are not in SkyComponent dataclass
             del sky_component_dict["id"]
+            del sky_component_dict["gsm_id"]
             del sky_component_dict["healpix_index"]
-            del sky_component_dict["version"]
-            del sky_component_dict["catalogue_name"]
             sky_components_dict[sky_component.id] = SkyComponentDataclass(**sky_component_dict)
 
-        return GlobalSkyModel(metadata=metadata_dict, components=sky_components_dict)
+        return GlobalSkyModel(
+            metadata=metadata_records[0].columns_to_dict(), components=sky_components_dict
+        )
 
     except Exception as e:
         logger.exception("Error querying GSM database: %s", e)
@@ -419,18 +427,19 @@ def _write_data(
         max_vector_len=5,  # For spectral index vectors
     )
 
-    local_model.set_header({"QUERY_CENTRE_RAJ2000_DEG": query_parameters.ra_deg})
-    local_model.set_header({"QUERY_CENTRE_DEJ2000_DEG": query_parameters.dec_deg})
-    local_model.set_header({"QUERY_RADIUS_DEG": query_parameters.fov_deg})
-    local_model.set_header({"QUERY_CATALOGUE_VERSION": query_parameters.version})
-    local_model.set_header({"QUERY_CATALOGUE_NAME": query_parameters.catalogue_name})
-    local_model.set_header({"CATALOGUE_VERSION": data.metadata.get("version", "unknown")})
-    local_model.set_header({"CATALOGUE_NAME": data.metadata.get("catalogue_name", "unknown")})
+    header = {
+        f"CATALOGUE_METADATA_{key}".upper(): value
+        for key, value in data.metadata.items()
+        if key not in ("staging", "upload_id", "id")
+    }
+    for key, value in query_parameters.__dict__.items():
+        if isinstance(value, dict):
+            for key2, val2 in value.items():
+                header[f"QUERY_{key}_{key2}".upper()] = val2
+        else:
+            header[f"QUERY_{key}".upper()] = value
 
-    for key, value in (
-        query_parameters.component_queries | query_parameters.metadata_queries
-    ).items():
-        local_model.set_header({f"QUERY_EXTRA_{key}": value})
+    local_model.set_header(header)
 
     # Populate the LocalSkyModel with data from GlobalSkyModel
     for row_idx, component in enumerate(data.components.values()):